+++ platforms/kia_soul/firmware/brake/tests/brake_controller/src/brake_controller_r0_test.ino
@@ -101,17 +101,17 @@ const byte PIN_PRR = 15;       // pressure rear right sensor
 
 
 // the following are guesses, these need to be debugged/researched
-const double ZERO_PRESSURE = 0.48;        // The voltage the sensors read when no pressure is present
-const double MIN_PACC = 2.3;              // minumum accumulator pressure to maintain
-const double MAX_PACC = 2.4;              // max accumulator pressure to maintain
-const double PEDAL_THRESH = 0.5;          // Pressure for pedal interference
+const float ZERO_PRESSURE = 0.48;        // The voltage the sensors read when no pressure is present
+const float MIN_PACC = 2.3;              // minumum accumulator pressure to maintain
+const float MAX_PACC = 2.4;              // max accumulator pressure to maintain
+const float PEDAL_THRESH = 0.5;          // Pressure for pedal interference
 
 int SLADutyMax,
     SLADutyMin,
     SLRDutyMax,
     SLRDutyMin;
 
-double pressure_req,
+float pressure_req,
        pressure,
        pressure_last,
        pressureRate_target,
@@ -684,7 +684,7 @@ void brakeUpdate()
 
     if( ret == PID_SUCCESS )
     {
-        double pressurePID_output = pidParams.control;
+        float pressurePID_output = pidParams.control;
 
         // constrain to min/max
         pressurePID_output = m_constrain(
